requirements_dev.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
kombu==5.2.3
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.26
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.5
pytz==2021.3
redis==4.1.2
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
